OS inpependent CMake Port

+++ src/Framework/Pipeline.cpp
@@ -1,4 +1,4 @@
-#include "LumenPCH.h"
+#include "../LumenPCH.h"
 #include "Pipeline.h"
 #include "VkUtils.h"
 
@@ -459,7 +459,7 @@ void Pipeline::create_update_template(const std::vector<Shader>& shaders,
 
 			default:
 				LUMEN_ERROR("Unimplemented descriptor type!");
-				return (unsigned long long)0;
+				return (size_t)0;
 		}
 	};
